requirements.txt
google-generativeai>=0.3.0

PyMuPDF>=1.23.0
Pillow>=10.0.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
fpdf-table>=0.1.0 